feat: final ui for history image
