+++ ui/src/features/project/pipelines/images.tsx
@@ -1,9 +1,9 @@
-import { faDocker } from '@fortawesome/free-brands-svg-icons';
-import { IconDefinition, faBook, faEyeSlash } from '@fortawesome/free-solid-svg-icons';
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { IconDefinition, faHistory, faEyeSlash } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Tooltip } from 'antd';
 import classNames from 'classnames';
-import { memo, useContext, useEffect, useMemo, useState } from 'react';
+import { memo, useContext, useMemo, useState, useEffect } from 'react';
 import { generatePath, useNavigate } from 'react-router-dom';
 import semver from 'semver';
 
@@ -13,12 +13,27 @@ import { TagMap } from '@ui/gen/api/service/v1alpha1/service_pb';
 import { Stage, Warehouse } from '@ui/gen/api/v1alpha1/generated_pb';
 import { useLocalStorage } from '@ui/utils/use-local-storage';
 
+type ProcessedTagMap = {
+  tags: Record<string, any>;
+};
+type ProcessedImages = Record<string, ProcessedTagMap>;
+
+const darkenColor = (hex: string, percent: number): string => {
+  if (!hex || !hex.startsWith('#') || hex.length !== 7) return '#4b5563';
+  const calculate = (c: string) =>
+    Math.max(0, Math.floor(parseInt(c, 16) * (1 - percent / 100)))
+      .toString(16)
+      .padStart(2, '0');
+  const r = calculate(hex.substring(1, 3));
+  const g = calculate(hex.substring(3, 5));
+  const b = calculate(hex.substring(5, 7));
+  return `#${r}${g}${b}`;
+};
+
 const findWarehousesForImageRepo = (repoURL: string, warehouses: Warehouse[]): string[] => {
   return warehouses
-    .filter((warehouse) =>
-      warehouse.spec?.subscriptions?.some((sub) => sub.image?.repoURL === repoURL)
-    )
-    .map((warehouse) => warehouse.metadata?.name || '')
+    .filter((w) => w.spec?.subscriptions?.some((s) => s.image?.repoURL === repoURL))
+    .map((w) => w.metadata?.name || '')
     .filter(Boolean);
 };
 
@@ -26,11 +41,7 @@ const findStagesForWarehouse = (warehouseName: string, stages: Stage[]): Set<str
   const reachableStages = new Set<string>();
   stages.forEach((stage) => {
     const stageName = stage.metadata?.name;
-    if (!stageName) return;
-    const requestsFromWarehouse = stage.spec?.requestedFreight?.some(
-      (req) => req.origin?.name === warehouseName
-    );
-    if (requestsFromWarehouse) {
+    if (stageName && stage.spec?.requestedFreight?.some((r) => r.origin?.name === warehouseName)) {
       reachableStages.add(stageName);
     }
   });
@@ -45,80 +56,131 @@ const getStagesForImage = (
   const warehousesForImage = findWarehousesForImageRepo(repoURL, warehouses);
   const allReachableStages = new Set<string>();
   warehousesForImage.forEach((warehouseName) => {
-    const stagesForWarehouse = findStagesForWarehouse(warehouseName, stages);
-    stagesForWarehouse.forEach((stageName) => allReachableStages.add(stageName));
+    findStagesForWarehouse(warehouseName, stages).forEach((stageName) =>
+      allReachableStages.add(stageName)
+    );
   });
   return allReachableStages;
 };
 
-const ImageTagRow = ({
-  tag,
-  stages,
-  imageStageMap,
-  projectName,
-  showHistory
-}: {
-  tag: string;
-  stages: Stage[];
-  imageStageMap: { [key: string]: number };
-  projectName: string;
-  showHistory: boolean;
-}) => {
-  const { stageColorMap } = useContext(ColorContext);
-  const navigate = useNavigate();
+const StageBox = memo(
+  ({
+    stageName,
+    order,
+    showHistory,
+    stageColorMap,
+    project
+  }: {
+    stageName: string;
+    order: number;
+    showHistory: boolean;
+    stageColorMap: Record<string, string>;
+    project: string;
+  }) => {
+    const navigate = useNavigate();
+    const isHighlighted = showHistory && order === 0;
+    const baseColor = stageColorMap[stageName] || '#6b7280';
 
-  return (
-    <div className='flex items-center mb-2'>
-      <Tooltip title={tag}>
-        <div className='mr-4 font-mono text-sm text-right w-20 truncate'>{tag}</div>
+    const tooltipTitle = showHistory
+      ? isHighlighted
+        ? `Most recent promotion: Currently in stage '${stageName}'`
+        : `Stage: '${stageName}' (Sequential promotion #${order + 1})`
+      : `Promoted to stage: '${stageName}'`;
+
+    const style = {
+      backgroundColor: baseColor,
+      opacity: showHistory && !isHighlighted ? 0.6 : 1,
+      border: isHighlighted ? `2px solid ${darkenColor(baseColor, 20)}` : '2px solid transparent',
+      boxShadow: isHighlighted ? '0 1px 3px rgba(0, 0, 0, 0.2)' : 'none'
+    };
+
+    return (
+      <Tooltip title={tooltipTitle}>
+        <div
+          onClick={() => navigate(generatePath(paths.stage, { name: project, stageName }))}
+          className='h-6 w-full rounded flex items-center justify-center cursor-pointer transition-all duration-300'
+          style={style}
+        >
+          {showHistory && (
+            <div className='flex items-center gap-0.5'>
+              <span className='text-white font-bold text-[10px] select-none'>#{order + 1}</span>
+            </div>
+          )}
+        </div>
       </Tooltip>
-      {stages.map((stage) => {
-        const stageName = stage.metadata?.name || '';
-        const cur = imageStageMap?.[stageName];
-        const len = stages.length;
+    );
+  }
+);
+StageBox.displayName = 'StageBox';
 
-        let opacity = 0;
-        if (cur !== undefined && !isNaN(cur)) {
-          opacity = len > 0 && showHistory ? 1 - cur / len : 1;
-        }
+const ImageTagRow = memo(
+  ({
+    tag,
+    imageStageMap,
+    dynamicStages,
+    showHistory,
+    project,
+    stageColorMap
+  }: {
+    tag: string;
+    imageStageMap: any;
+    dynamicStages: string[];
+    showHistory: boolean;
+    project: string;
+    stageColorMap: Record<string, string>;
+  }) => {
+    const sequentialOrderMap = useMemo(() => {
+      if (!showHistory) return {};
 
-        return (
-          <Tooltip key={stageName} title={stageName}>
-            <div
-              className={classNames('mr-2 transition-all duration-200', {
-                'cursor-pointer': cur !== undefined && cur >= 0
-              })}
-              style={{
-                borderRadius: '5px',
-                height: '30px',
-                width: '30px',
-                opacity,
-                backgroundColor:
-                  (showHistory && cur !== undefined && cur >= 0) || cur === 0
-                    ? stageColorMap?.[stageName]
-                    : '#e5e7eb',
-                border:
-                  cur === 0 && showHistory
-                    ? '3px solid rgba(255, 255, 255, 0.4)'
-                    : '3px solid transparent'
-              }}
-              onClick={() => {
-                if (cur !== undefined) {
-                  navigate(
-                    generatePath(paths.stage, {
-                      name: projectName,
-                      stageName: stageName
-                    })
-                  );
-                }
-              }}
-            />
+      const stagesWithOrder = Object.entries(imageStageMap.stages || {})
+        .filter(([stageName]) => dynamicStages.includes(stageName))
+        .sort(([, a], [, b]) => (a as number) - (b as number));
+
+      const orderMap: Record<string, number> = {};
+      stagesWithOrder.forEach(([stageName], index) => {
+        orderMap[stageName] = index;
+      });
+
+      return orderMap;
+    }, [imageStageMap.stages, dynamicStages, showHistory]);
+
+    return (
+      <tr className='hover:bg-gray-50/70'>
+        <td className='sticky left-0 bg-white hover:bg-gray-50/70 px-1 py-0.5 border-b border-gray-200 z-20'>
+          <Tooltip title={`Image Tag: ${tag}`}>
+            <div className='font-mono text-xs font-semibold truncate'>{tag}</div>
           </Tooltip>
-        );
-      })}
-    </div>
-  );
-};
+        </td>
+        {dynamicStages.map((stageName) => {
+          const originalOrder = imageStageMap.stages?.[stageName];
+          const hasImage = originalOrder !== undefined;
+          const sequentialOrder = sequentialOrderMap[stageName];
+
+          return (
+            <td key={stageName} className='px-1 py-1 border-b border-gray-200'>
+              <div className='flex justify-center'>
+                {hasImage ? (
+                  <StageBox
+                    stageName={stageName}
+                    order={sequentialOrder}
+                    showHistory={showHistory}
+                    stageColorMap={stageColorMap}
+                    project={project}
+                  />
+                ) : (
+                  <div className='w-full h-6 bg-gray-100 rounded text-xs flex items-center justify-center text-gray-400'>
+                    -
+                  </div>
+                )}
+              </div>
+            </td>
+          );
+        })}
+      </tr>
+    );
+  }
+);
+ImageTagRow.displayName = 'ImageTagRow';
 
 const HeaderButton = memo(
   ({
@@ -136,8 +198,8 @@ const HeaderButton = memo(
       <button
         onClick={onClick}
         className={classNames(
-          'p-2 w-7 h-7 flex items-center justify-center rounded-md hover:bg-gray-300 transition-colors',
-          selected ? 'bg-blue-100 text-blue-600' : 'text-gray-400 hover:text-gray-500'
+          'p-2 w-7 h-7 flex items-center justify-center rounded-md hover:bg-gray-200 transition-colors',
+          selected ? 'bg-blue-100 text-blue-600' : 'text-gray-500 hover:text-gray-700'
         )}
       >
         <FontAwesomeIcon icon={icon} />
@@ -147,122 +209,181 @@ const HeaderButton = memo(
 );
 HeaderButton.displayName = 'HeaderButton';
 
-const Select = memo(
-  ({
-    value,
-    onChange,
-    options
-  }: {
-    value: string;
-    onChange: (value: string) => void;
-    options: { label?: string; value: string }[];
-  }) => (
-    <select
-      className='block border-none w-full text-gray-600 appearance-none p-2 bg-gray-200 focus:outline-none focus:ring-2 focus:ring-blue-400 rounded-md'
-      value={value}
-      onChange={(e) => onChange(e.target.value)}
-    >
-      {options.map((option) => (
-        <option value={option.value} key={option.value}>
-          {option.label}
-        </option>
-      ))}
-    </select>
-  )
-);
-Select.displayName = 'Select';
+interface ImagesProps {
+  hide: () => void;
+  images: Record<string, TagMap>;
+  project: string;
+  stages: Stage[];
+  warehouses: Warehouse[];
+}
 
-export const Images = memo(
-  ({
-    project,
-    stages,
-    hide,
-    images,
-    warehouses
-  }: {
-    project: string;
-    stages: Stage[];
-    hide: () => void;
-    images: { [key: string]: TagMap };
-    warehouses: Warehouse[];
-  }) => {
-    const imageRepos = useMemo(() => Object.keys(images || {}), [images]);
-    const [imageURL, setImageURL] = useState(imageRepos[0] || '');
-    const [showHistory, setShowHistory] = useLocalStorage(`${project}-show-history`, true);
+export const Images = memo<ImagesProps>(({ hide, images, project, stages, warehouses }) => {
+  const { stageColorMap } = useContext(ColorContext);
+  const [showHistory, setShowHistory] = useLocalStorage('images-dynamic-grid-show-history', true);
 
-    useEffect(() => {
-      if (!imageRepos.includes(imageURL)) {
-        setImageURL(imageRepos[0] || '');
+  const filteredImages: ProcessedImages = useMemo(() => {
+    const result: ProcessedImages = {};
+    Object.entries(images).forEach(([repoURL, tagMap]) => {
+      const stagesForImage = getStagesForImage(repoURL, warehouses, stages);
+      const filteredTagMap: ProcessedTagMap = { tags: {} };
+      Object.entries(tagMap.tags || {}).forEach(([tag, imageStageMap]) => {
+        const filteredStages: Record<string, number> = {};
+        Object.entries(imageStageMap.stages || {}).forEach(([stageName, order]) => {
+          if (stagesForImage.has(stageName)) {
+            filteredStages[stageName] = order;
+          }
+        });
+        if (Object.keys(filteredStages).length > 0) {
+          filteredTagMap.tags[tag] = { ...imageStageMap, stages: filteredStages };
+        }
+      });
+      if (Object.keys(filteredTagMap.tags).length > 0) {
+        result[repoURL] = filteredTagMap;
       }
-    }, [imageRepos, imageURL]);
+    });
+    return result;
+  }, [images, warehouses, stages]);
 
-    const relevantStages = useMemo(
-      () => (imageURL ? getStagesForImage(imageURL, warehouses, stages) : new Set<string>()),
-      [imageURL, warehouses, stages]
-    );
+  const repoURLs = useMemo(() => Object.keys(filteredImages), [filteredImages]);
+  const [selectedRepoURL, setSelectedRepoURL] = useState<string>('');
 
-    const filteredStages = useMemo(
-      () => stages.filter((stage) => relevantStages.has(stage.metadata?.name || '')),
-      [stages, relevantStages]
-    );
+  useEffect(() => {
+    if (repoURLs.length > 0 && !repoURLs.includes(selectedRepoURL)) {
+      setSelectedRepoURL(repoURLs[0]);
+    } else if (repoURLs.length === 0) {
+      setSelectedRepoURL('');
+    }
+  }, [repoURLs, selectedRepoURL]);
 
-    const curImage = images[imageURL];
-    const sortedTags = useMemo(() => {
-      return curImage?.tags
-        ? Object.keys(curImage.tags || {}).sort((a, b) => {
-            try {
-              return semver.compare(b, a);
-            } catch (e) {
-              return b.localeCompare(a);
-            }
-          })
-        : [];
-    }, [curImage]);
+  // NEW: Derive data based on the selection
+  const selectedImageData = useMemo(
+    () => (selectedRepoURL ? filteredImages[selectedRepoURL] : undefined),
+    [selectedRepoURL, filteredImages]
+  );
 
-    return (
-      <div className='text-gray-600 text-sm bg-gray-100 pb-4 rounded-md'>
-        <h3 className='bg-gray-200 px-4 py-2 flex items-center text-sm text-gray-500'>
-          <FontAwesomeIcon icon={faDocker} className='mr-2' /> IMAGES
-          <div className='ml-auto flex items-center gap-2'>
-            <HeaderButton
-              onClick={() => setShowHistory(!showHistory)}
-              icon={faBook}
-              selected={showHistory}
-              title={showHistory ? 'Hide promotion history' : 'Show promotion history'}
-            />
-            <HeaderButton onClick={hide} icon={faEyeSlash} title='Hide images panel' />
-          </div>
-        </h3>
-        <div className='p-4 overflow-y-auto max-h-[356px]'>
-          {imageRepos.length > 0 && imageURL ? (
-            <>
-              <div className='mb-4'>
-                <Select
-                  value={imageURL}
-                  onChange={setImageURL}
-                  options={imageRepos.map((image) => ({
-                    label: image.split('/').pop(),
-                    value: image
-                  }))}
-                />
-              </div>
-              {sortedTags.map((tag) => (
-                <ImageTagRow
-                  key={tag}
-                  projectName={project}
-                  tag={tag}
-                  stages={filteredStages}
-                  imageStageMap={curImage.tags[tag]?.stages || {}}
-                  showHistory={showHistory}
-                />
-              ))}
-            </>
-          ) : (
-            <p className='text-center text-gray-500 py-8'>No images available for this project.</p>
+  const dynamicStages = useMemo(() => {
+    if (!selectedImageData) return [];
+    const stageSet = new Set<string>();
+    Object.values(selectedImageData.tags).forEach((imageStageMap) => {
+      Object.keys(imageStageMap.stages || {}).forEach((stageName) => stageSet.add(stageName));
+    });
+    return Array.from(stageSet).sort();
+  }, [selectedImageData]);
+
+  const allTags = useMemo(() => {
+    if (!selectedImageData) return [];
+    return Object.keys(selectedImageData.tags).sort((a, b) => {
+      try {
+        return semver.compare(b, a);
+      } catch {
+        return b.localeCompare(a);
+      }
+    });
+  }, [selectedImageData]);
+
+  return (
+    <div className='bg-white rounded-lg shadow-xl border border-gray-200/75 p-3'>
+      <div className='flex items-center justify-between mb-2'>
+        <div className='flex items-baseline gap-2'>
+          <h3 className='text-base font-semibold text-gray-800'>Images</h3>
+          {selectedRepoURL && (
+            <p className='text-sm text-gray-600 truncate'>{selectedRepoURL.split('/').pop()}</p>
           )}
         </div>
+        <div className='flex items-center gap-1'>
+          <HeaderButton
+            onClick={() => setShowHistory(!showHistory)}
+            selected={showHistory}
+            icon={faHistory}
+            title={showHistory ? 'Hide promotion history' : 'Show promotion history'}
+          />
+          <HeaderButton onClick={hide} icon={faEyeSlash} title='Hide panel' />
+        </div>
       </div>
-    );
-  }
-);
+      {showHistory && (
+        <div className='mb-1 text-xs text-gray-500 bg-gray-50 px-2 py-0.5 rounded'>
+          Numbers show sequential promotion order (#1 = most recent)
+        </div>
+      )}
+
+      {repoURLs.length > 1 && (
+        <div className='mb-3'>
+          <select
+            value={selectedRepoURL}
+            onChange={(e) => setSelectedRepoURL(e.target.value)}
+            className='block border-gray-300 w-full text-gray-800 appearance-none p-2 bg-white rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500'
+          >
+            {repoURLs.map((repo) => (
+              <option value={repo} key={repo}>
+                {repo.split('/').pop() || repo}
+              </option>
+            ))}
+          </select>
+        </div>
+      )}
+
+      <div className='overflow-x-auto max-h-[356px] relative z-10'>
+        <table className='text-sm border-collapse w-full' style={{ tableLayout: 'fixed' }}>
+          <colgroup>
+            <col style={{ width: '120px' }} />
+            {dynamicStages.map((stageName) => (
+              <col key={stageName} style={{ width: '80px' }} />
+            ))}
+          </colgroup>
+          <thead>
+            <tr className='bg-gray-50'>
+              <th className='sticky left-0 bg-gray-50 text-left px-1 py-0.5 border-b font-semibold text-gray-600 z-20'>
+                Tag
+              </th>
+              {dynamicStages.map((stageName) => (
+                <th
+                  key={stageName}
+                  className='px-1 py-1 border-b font-semibold text-gray-600 text-center'
+                >
+                  <Tooltip title={stageName}>
+                    <div className='flex items-center justify-center gap-1'>
+                      <div
+                        className='w-2.5 h-2.5 rounded-full flex-shrink-0'
+                        style={{ backgroundColor: stageColorMap[stageName] }}
+                      />
+                      <span className='text-xs font-medium text-gray-700 truncate max-w-[60px]'>
+                        {stageName}
+                      </span>
+                    </div>
+                  </Tooltip>
+                </th>
+              ))}
+            </tr>
+          </thead>
+          <tbody>
+            {allTags.length > 0 ? (
+              allTags.map((tag) => {
+                const imageStageMap = selectedImageData?.tags[tag];
+                if (!imageStageMap) return null;
+                return (
+                  <ImageTagRow
+                    key={tag}
+                    tag={tag}
+                    imageStageMap={imageStageMap}
+                    dynamicStages={dynamicStages}
+                    showHistory={showHistory}
+                    project={project}
+                    stageColorMap={stageColorMap}
+                  />
+                );
+              })
+            ) : (
+              <tr>
+                <td colSpan={dynamicStages.length + 1} className='text-center text-gray-500 py-12'>
+                  {repoURLs.length > 0 ? 'Select an image' : 'No images found'}
+                </td>
+              </tr>
+            )}
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+});
+
 Images.displayName = 'Images';
